Fix a race condition in getting the trunk name via its mac.
During setting up a branch interface, it can have the same mac address as the trunk for a short period of time, and we may end up picking the branch's name when trying to get the trunk name by mac. Fixing it by picking the shortest name available since branch name is strictly longer than trunk.